requirements.txt
Flask
Flask-Cors
python-dotenv
gunicorn
google-generativeai
